+++ packages/insomnia/src/ui/components/editors/__tests__/environment-editor.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from '@jest/globals';
 
 import { NUNJUCKS_TEMPLATE_GLOBAL_PROPERTY_NAME } from '../../../../templating';
-import { checkNestedKeys, ensureKeyIsValid } from '../environment-editor';
+import { checkNestedKeys, ensureKeyIsValid } from '../environment-utils';
 
 describe('ensureKeyIsValid()', () => {
   it.each([
