+++ packages/insomnia/src/ui/components/editors/body/graph-ql-editor.tsx
@@ -448,6 +448,7 @@ export const GraphQLEditor: FC<Props> = ({
   if (schema) {
     graphqlOptions = {
       hintOptions: {
+        schema,
         completeSingle: false,
       },
       infoOptions: {
